Users can set url_keys for pages
To make urls more memorable, pleasing, the concept of url_keys is introduced. Each page has a url_key which is the url key that can be set by the user. Forms need updating to support this, and validations are added.

Also the actual show page itself is introduced. Previously this functionality was all in Backbone. The user can choose which pages they want on their website root menu (about me, projects, resume, links, etc) and these have their own show page (currently in the welcome controller that should probably be renamed.