Easier to Understand Errors from Page save
One create or update, if errors were present that prohib saving the app would render index and then the user sees the top of the page. If they have many pages this is confusing. Instead, errors on create render new (with form filled out and flash errors are adjacent to it. On update error we render edit.