One Photo is loaded in large form on the show page
The rest of the photos on a project show page are thumbnail size. The one prominently displayed is specified by the aprams. A model called View::DisplayPhoto is used to spit out the correct photo. This helper takes in the params and then chooses the correct photo url. If no params are given, the default is the avatar attached to the projet.
This abstraction doesn't feel great, but I am enjoying copying Sandy's (Luke's mom) art website. I like the thumbnail layout and the one prominent picture.