Edit Projects From Edit Page
This also allows photos on a project to be deleted. A thumb image next to the delete button shows which phot will be deleted. A photos_controller_spec is added and the destroy method is tested. An auth_helper file is added to DRY up the stubbing of authenticated users across controller specs. Photos can be added to the project from the edit project page as well. This uses the photos controller rather than projects controller. The method in the photos controller had to be modified and tested (cropping is no longer possible here, but may be added back in later)