Associate users with models
Originally this app was only meant to have one user (an admin like
function) that could edit, upload, and delete. Since this project will
now be a portfolio for multiple people with multiple domains hitting the
same IP address, users will be used to map to content.
A user will have and be able to edit their projects, photos, and custom
pages.

* pages associated with user
  - remove uniqueness constraints on url keys and titles and indexes
    that required uniqueness
  - rails now validates uniqueness within a scope (user_id)
  - Sequences are used for factory girl model of page so that uniqueness
    constraint is not violated in tests.
* Users map to a domain
  - This can be checked in the request (request.host)
  - Helper method "user_by_domain" determines what is shown. However
    only the "current_user" (helper method from devise) can edit their
    own content.
  - Figure out creation of users later (probably manual)
  - Need a test user for the domain ("test.host")
* Projects associated with user
  - Project index page viewable by anyone
  - Other pages protected (current_user)