Rework styling on welcome index and show
The root of the application is the welcome index page. This shows a menu
of "page" titles that link to each user created page. The styling here
was not very thought out. This uses bootstrap a bit better and will
provide a better mobile experience as well.
My phone is being a difficult-diva at the moment so I will not make
mobile changes in this commit.

* One important difference is that links are easier to click on the menu
  index since each link is a block that spans the whole page
  (nav-stacked class bootstrap for <ul>)
* Created a mixin to DRY things up and reuse later for adjusting size on
  mobile. It's definitely not the greatest abstraction/use of mixins,
  but it was fun.