Domain determined by user
Previously this application was many separate heroku apps and users were
only used as admin (one user per app). At that time the title was
an environment variable. Now the title should be specific to the user
tied to the host/domain.

The user_by_domain is also memoized since this is often queried. Not
sure how efficient ActiveRecord caching is.

Fixes failing spec. Feature spec was looking for the page title, but
this part is only shown as an editable field in a form.