Ensure user signs in at correct domain
It is confusing to visit the sign in path under the wrong host, sign in
and then try to edit things. Authentication prevents users editing the
pages and projects of other users, but one wastes time trying without a
helpful error message.

This adds an error message on an attempt at sign by checking whether the
users email matches the user_by_domain email. A user must visit their
own domain to log in. This required a custom sessions controller that
only handles checking the email and user_by_domain before calling
`super`

This also prevents a signed in user from visiting the pages index at
`/pages` under the wrong domain.

Tests were also added around uniqueness constraints for page attributes
(url_key and order). This was working fine, but there are no tests for
it. The problems were related to the above problems with logging in and
using the wrong domain.