Users can only sign in under correct domain
Test failed because behavior changed in last commit to prevent this
problem:

This tested that users cannot edit pages that are not their own.
However, one cannot easily sign in under a different domain in testing
(one could try modifying the host or stubbing that), but it seemed
easiest just to change the users domain after sign in.