Adds live preview functionality
A user can enter text into the form to create a page, and they will see
a live preview of what the markdown will look like. This is copied from:
https://github.com/sophomoric/secret/commit/9713d025edfaa3fa1924d555ba522e9b1a6bed08

Since the number of users on this website is limited, having requests
made for every keyup event is not a big load on the server.

Using JS in the feature spec comes with some complications. Since this
app has checks for a user editing pages out of their domain (host), the
app_host during the test is crucial at certain points. When `js: true`
is set, a JS driver is being used. Capybara starts a server on another
thread (the default driver does not do this). To make the spec pass I
set the user's domain property to the host server that Capybara is
running. There may be a better way than hard coding this but for now it
passes.

Database cleaner also required tweaking to prevent the user being
deleted during the `js: true` feature specs that sign in through the
"back door"

Edit each page on its own page.