Stylistic changes
Editing pages is an option from the user menu, but adding a new page and
swapping the order of pages are require links somewhere. These options
will be links once this option is clicked. They appear green so they
stand out, and it is clear they are not pages to be edited.

* Also fix styling bug where the new page form had too much margin above
  it.