Display user photos on the page edit form
This allows users to insert an image into a page by clicking the
thumbnail of the image. The image tag is appended to the input value of
the body.

The current downside is that this forces another request to the server
for markdown, increasing the load on the server and the client. I am
chosing to do this rather than append the image to the preview section
because I believe the next step will involve allowing the user to
specify the size of the image, and I will end up wanting to trigger the
"keyup" event anyway for simplicity.