Users can swap the order of pages on the root menu
Currently each page has a persisted attribute `order` that determines
where it appears on the root menu. Now users can modify this attribute
after a page is created. The unique constraint on page.order requires
a number of calls to the database to swap the order of pages. First, all
page orders are set to -1 (skipping rails validations). Then pages are
ordered based on user input. This means every page is re-saved even if
the order remained the same.

The abovementioned process is wrapped in a transaction so the original
state can be preserved if other errors occur.

Possible Improvements:
* Optimize later for less saving
* `include ActiveModel::Validations` in `OrderSwap` model to record and
  display better errors for the user
* Refactor domain setting into factory
* Nest `order_swap` under `pages` to be more explicit

Random Notes:
TIL: since tests are wrapped in a transaction:
http://stackoverflow.com/questions/21340686/rails-transaction-not-rolling-back
need to pass the `requires_new: true` since this is a nested transaction
in when running a test:
http://api.rubyonrails.org/classes/ActiveRecord/Transactions/ClassMethods.html

"If your commit message is more than 15 words you should probably just
write it down and read it to your therapist."
https://twitter.com/iamdevloper/status/657130347109351424