Prevent hidden pages from being scrolled through
This allows pages to exist that will not be found unless the url_key is
known or guessed. Previously, the page could be hidden from the root
url, but if one clicked "next" or "back" (</>) the page would be found.

Use case: I put up a weird page at `/secret-page` and I do not want it
to be found if someone visits the website and scrolls through pages by
clicking next: ">"